requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pymongo>=4.6.0
python-multipart>=0.0.6
httpx>=0.25.0
pydantic>=2.4.2
python-dotenv>=1.0.0
motor>=3.3.1
pydantic-settings>=2.0.0
email-validator>=2.0.0 